% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FScolar.R
\name{FScolar}
\alias{FScolar}
\title{Cola dois paramentros de entrada}
\usage{
FScolar(a, b)
}
\arguments{
\item{a}{primeiro parametro}

\item{b}{segundo parametro}
}
\value{
o paste dos dois parametros
}
\description{
Recebe dois parametros e colam os mesmos numa unica saída
}
